+++ packages/sit-onyx/src/components/TestInput/TestInput.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import { injectI18n } from "@/i18n";
-import { areObjectsFlatEqual } from "@/utils/comparator";
+import { areObjectsFlatEqual } from "@/utils/objects";
 import { getFirstInvalidType, transformValidityStateToObject } from "@/utils/forms";
 import { computed, ref, toRefs, watch } from "vue";
 import type { InputType } from "../OnyxInput/types";
