+++ packages/sit-onyx/src/utils/objects.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "vitest";
-import { areObjectsFlatEqual } from "./comparator";
+import { areObjectsFlatEqual } from "./objects";
 
 const referenceObj = { a: 42, b: "foo", c: null, d: true };
